# src/api/rest/docs/swagger.yaml
basePath: /tumblebug
definitions:
...
  infra.JSONResult:
    type: object
...
  infra.RestGetAllK8sClusterResponse:
    properties:
      cluster:
        items:
          $ref: '#/definitions/model.TbK8sClusterInfo'
        type: array
    type: object
...
  model.IdList:
    properties:
      output:
        items:
          type: string
        type: array
    type: object
...
  model.K8sClusterNodeGroupsOnCreation:
    properties:
      result:
        example: "true"
        type: string
    type: object
...
  model.K8sClusterNodeImageDetailAvailable:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  model.TbChangeK8sNodeGroupAutoscaleSizeReq:
    properties:
      desiredNodeSize:
        example: "1"
        type: string
      maxNodeSize:
        example: "3"
        type: string
      minNodeSize:
        example: "1"
        type: string
    type: object
...
  model.TbK8sClusterReq:
    properties:
      connectionName:
        description: Namespace      string `json:"namespace" validate:"required" example:"default"`
        example: alibaba-ap-northeast-2
        type: string
      cspK8sClusterId:
        description: |-
          Fields for "Register existing K8sCluster" feature
          @description CspK8sClusterId is required to register a k8s cluster from CSP (option=register)
        example: required when option is register
        type: string
      description:
        example: My K8sCluster
        type: string
      id:
        description: (1) K8sCluster Info
        example: k8scluster-01
        type: string
      k8sNodeGroupList:
        description: (3) NodeGroupInfo List
        items:
          $ref: '#/definitions/model.TbK8sNodeGroupReq'
        type: array
      securityGroupIds:
        example:
        - sg-01
        items:
          type: string
        type: array
      subnetIds:
        example:
        - subnet-01
        items:
          type: string
        type: array
      vNetId:
        description: (2) Network Info
        example: vpc-01
        type: string
      version:
        example: 1.30.1-aliyun.1
        type: string
    required:
    - connectionName
    - id
    - securityGroupIds
    - subnetIds
    - vNetId
    type: object
...
  model.TbSetK8sNodeGroupAutoscalingReq:
    properties:
      onAutoScaling:
        example: "true"
        type: string
    type: object
  model.TbSetK8sNodeGroupAutoscalingRes:
    properties:
      result:
        example: "true"
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cb-tumblebug/issues/new/choose
  description: CB-Tumblebug is an open source system for managing multi-cloud infrastructure
    consisting of resources from multiple cloud service providers. (Cloud-Barista)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cloud-barista/cb-tumblebug/blob/main/README.md
  title: CB-Tumblebug REST API
  version: latest
paths:
...
  /availableK8sClusterNodeImage:
    get:
      consumes:
      - application/json
      description: Get available kubernetes cluster node image
      operationId: GetAvailableK8sClusterNodeImage
      parameters:
      - description: Name of the CSP to retrieve
        in: query
        name: providerName
        required: true
        type: string
      - description: Name of region to retrieve
        in: query
        name: regionName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.K8sClusterNodeImageDetailAvailable'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get available kubernetes cluster node image
      tags:
      - '[Kubernetes] Cluster Management'
...
  /checkNodeGroupsOnK8sCreation:
    get:
      consumes:
      - application/json
      description: Check whether nodegroups are required during the k8scluster creation
      operationId: CheckNodeGroupsOnK8sCreation
      parameters:
      - description: Name of the CSP to retrieve
        in: query
        name: providerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.K8sClusterNodeGroupsOnCreation'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Check whether nodegroups are required during the k8scluster creation
      tags:
      - '[Kubernetes] Cluster Management'
...
  /ns/{nsId}/k8scluster:
    delete:
      consumes:
      - application/json
      description: Delete all K8sClusters
      operationId: DeleteAllK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: ""
        description: Delete resources containing matched ID-substring only
        in: query
        name: match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IdList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete all K8sClusters
      tags:
      - '[Kubernetes] Cluster Management'
    get:
      consumes:
      - application/json
      description: List all K8sClusters or K8sClusters' ID
      operationId: GetAllK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Option
        enum:
        - id
        in: query
        name: option
        type: string
      - description: 'Field key for filtering (ex: cspK8sClusterName)'
        in: query
        name: filterKey
        type: string
      - description: 'Field value for filtering (ex: default-alibaba-ap-northeast-2-vpc)'
        in: query
        name: filterVal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/infra.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/infra.RestGetAllK8sClusterResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all K8sClusters or K8sClusters' ID
      tags:
      - '[Kubernetes] Cluster Management'
    post:
      consumes:
      - application/json
      description: Create K8sCluster<br>Find details from https://github.com/cloud-barista/cb-tumblebug/discussions/1614
      operationId: PostK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: 'Option: [required params for register] connectionName, name,
          cspK8sClusterId'
        enum:
        - register
        in: query
        name: option
        type: string
      - description: Details of the K8sCluster object
        in: body
        name: k8sClusterReq
        required: true
        schema:
          $ref: '#/definitions/model.TbK8sClusterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbK8sClusterInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Create K8sCluster
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}:
    delete:
      consumes:
      - application/json
      description: Delete K8sCluster
      operationId: DeleteK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete K8sCluster
      tags:
      - '[Kubernetes] Cluster Management'
    get:
      consumes:
      - application/json
      description: Get K8sCluster
      operationId: GetK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbK8sClusterInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get K8sCluster
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}/k8snodegroup:
    post:
      consumes:
      - application/json
      description: Add a K8sNodeGroup
      operationId: PostK8sNodeGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      - description: Details of the K8sNodeGroup object
        in: body
        name: k8sNodeGroupReq
        required: true
        schema:
          $ref: '#/definitions/model.TbK8sNodeGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbK8sClusterInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Add a K8sNodeGroup
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}/k8snodegroup/{k8sNodeGroupName}:
    delete:
      consumes:
      - application/json
      description: Remove a K8sNodeGroup
      operationId: DeleteK8sNodeGroup
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      - default: ng-01
        description: K8sNodeGroup Name
        in: path
        name: k8sNodeGroupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Remove a K8sNodeGroup
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}/k8snodegroup/{k8sNodeGroupName}/autoscalesize:
    put:
      consumes:
      - application/json
      description: Change a K8sNodeGroup's Autoscale Size
      operationId: PutChangeK8sNodeGroupAutoscaleSize
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      - default: ng-01
        description: K8sNodeGroup Name
        in: path
        name: k8sNodeGroupName
        required: true
        type: string
      - description: Details of the TbChangeK8sNodeGroupAutoscaleSizeReq object
        in: body
        name: changeK8sNodeGroupAutoscaleSizeReq
        required: true
        schema:
          $ref: '#/definitions/model.TbChangeK8sNodeGroupAutoscaleSizeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbChangeK8sNodeGroupAutoscaleSizeRes'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Change a K8sNodeGroup's Autoscale Size
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}/k8snodegroup/{k8sNodeGroupName}/onautoscaling:
    put:
      consumes:
      - application/json
      description: Set a K8sNodeGroup's Autoscaling On/Off
      operationId: PutSetK8sNodeGroupAutoscaling
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      - default: ng-01
        description: K8sNodeGroup Name
        in: path
        name: k8sNodeGroupName
        required: true
        type: string
      - description: Details of the TbSetK8sNodeGroupAutoscalingReq object
        in: body
        name: setK8sNodeGroupAutoscalingReq
        required: true
        schema:
          $ref: '#/definitions/model.TbSetK8sNodeGroupAutoscalingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbSetK8sNodeGroupAutoscalingRes'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Set a K8sNodeGroup's Autoscaling On/Off
      tags:
      - '[Kubernetes] Cluster Management'
  /ns/{nsId}/k8scluster/{k8sClusterId}/upgrade:
    put:
      consumes:
      - application/json
      description: Upgrade a K8sCluster's version
      operationId: PutUpgradeK8sCluster
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: k8scluster-01
        description: K8sCluster ID
        in: path
        name: k8sClusterId
        required: true
        type: string
      - description: Details of the TbUpgradeK8sClusterReq object
        in: body
        name: upgradeK8sClusterReq
        required: true
        schema:
          $ref: '#/definitions/model.TbUpgradeK8sClusterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Upgrade a K8sCluster's version
      tags:
      - '[Kubernetes] Cluster Management'
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token ([TBD] Get token
      in http://xxx.xxx.xxx.xxx:xxx/auth)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"